+++ packages/interactivity/src/index.ts
@@ -14,6 +14,7 @@ import { directive } from './hooks';
 import { getNamespace } from './namespaces';
 import { parseServerData, populateServerData } from './store';
 import { proxifyState } from './proxies';
+import { deepReadOnly, navigationSignal } from './utils';
 
 export {
 	store,
@@ -40,7 +41,9 @@ export { useState, useRef } from 'preact/hooks';
 const requiredConsent =
 	'I acknowledge that using private APIs means my theme or plugin will inevitably break in the next version of WordPress.';
 
-export const privateApis = ( lock ): any => {
+export const privateApis = (
+	lock: 'I acknowledge that using private APIs means my theme or plugin will inevitably break in the next version of WordPress.'
+): any => {
 	if ( lock === requiredConsent ) {
 		return {
 			getRegionRootFragment,
@@ -56,6 +59,8 @@ export const privateApis = ( lock ): any => {
 			populateServerData,
 			batch,
 			routerRegions,
+			deepReadOnly,
+			navigationSignal,
 		};
 	}